处理当前用户没有绑定默认微管理账号的异常情况

+++ MyCircle/MCMicroManagerViewController.m
@@ -180,8 +180,15 @@
     //获取当前账号所有功能模块
     DLog(@"--------didFinsishGetMMAccount-----------");
     DLog(@"微管理获取当前账号所有功能模块");
-    [[MCMicroManagerDAO sharedManager] deleteAll];
-    [[MCMicroManagerConfigHandler sharedInstance] getCodeByUserCode:mmAccount.userCode acctId:mmAccount.acctId];
+    if (！mmAccount.acctId) {
+        //当前用户没有绑定默认微管理账号
+        UIAlertView *alert = [UIAlertView alloc] initWithTitle@"错误" message:@"当前用户没有绑定默认微管理账号" delegate:nil cancelButtonTitle:@"确定" otherButtonTitles:nil];
+        [alert show];
+    }
+    else {
+        [[MCMicroManagerDAO sharedManager] deleteAll];
+        [[MCMicroManagerConfigHandler sharedInstance] getCodeByUserCode:mmAccount.userCode acctId:mmAccount.acctId];
+    }
 }
 
 - (void)didFinishGetMicroManagerWidget
